CORE-17065: Ensure the locally hosted identities check gets the right identity as the destination
